refactor: Separate document test suites out.
