+++ tests/plugins/document/suite.go
@@ -12,14 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package document
+package document_suite
 
 import (
-	"fmt"
-
 	"github.com/nitric-dev/membrane/pkg/plugins/document"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 )
 
 // Simple 'users' collection test data
@@ -169,7 +165,7 @@ var Customer1 = Customer{
 			},
 			Content: map[string]interface{}{
 				"title": "Good review",
-				"stars":      "5",
+				"stars": "5",
 			},
 		},
 	},
@@ -326,760 +322,3 @@ func LoadItemsData(docPlugin document.DocumentService) {
 }
 
 // Unit Test Functions --------------------------------------------------------
-
-func GetTests(docPlugin document.DocumentService) {
-	Context("Get", func() {
-		When("Blank key.Collection.Name", func() {
-			It("Should return error", func() {
-				key := document.Key{Id: "1"}
-				_, err := docPlugin.Get(&key)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Blank key.Id", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}}
-				_, err := docPlugin.Get(&key)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid Get", func() {
-			It("Should get item successfully", func() {
-				docPlugin.Set(&UserKey1, UserItem1)
-
-				doc, err := docPlugin.Get(&UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Key).To(Equal(&UserKey1))
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem1["email"]))
-			})
-		})
-		When("Valid Sub Collection Get", func() {
-			It("Should store item successfully", func() {
-				docPlugin.Set(&Customer1.Orders[0].Key, Customer1.Orders[0].Content)
-
-				doc, err := docPlugin.Get(&Customer1.Orders[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Key).To(Equal(&Customer1.Orders[0].Key))
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Orders[0].Content))
-			})
-		})
-		When("Document Doesn't Exist", func() {
-			It("Should return NotFound error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "items"}, Id: "not-exist"}
-				doc, err := docPlugin.Get(&key)
-				Expect(doc).To(BeNil())
-				Expect(err).Should(HaveOccurred())
-				Expect(err.Error()).To(ContainSubstring("not found"))
-			})
-		})
-		When("Valid Collection Get when there is a Sub Collection", func() {
-			It("Should store item successfully", func() {
-				docPlugin.Set(&Customer1.Key, Customer1.Content)
-
-				doc, err := docPlugin.Get(&Customer1.Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Key).To(Equal(&Customer1.Key))
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Content))
-			})
-		})
-	})
-}
-
-func SetTests(docPlugin document.DocumentService) {
-	Context("Set", func() {
-		When("Blank key.Collection.Name", func() {
-			It("Should return error", func() {
-				key := document.Key{Id: "1"}
-				err := docPlugin.Set(&key, UserItem1)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Blank key.Id", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}}
-				err := docPlugin.Set(&key, UserItem1)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Nil item map", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}, Id: "1"}
-				err := docPlugin.Set(&key, nil)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid New Set", func() {
-			It("Should store new item successfully", func() {
-				err := docPlugin.Set(&UserKey1, UserItem1)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(&UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem1["email"]))
-			})
-		})
-		When("Valid Update Set", func() {
-			It("Should update existing item successfully", func() {
-				err := docPlugin.Set(&UserKey1, UserItem1)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(&UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem1["email"]))
-
-				err = docPlugin.Set(&UserKey1, UserItem2)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err = docPlugin.Get(&UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem2["email"]))
-			})
-		})
-		When("Valid Sub Collection Set", func() {
-			It("Should store item successfully", func() {
-				err := docPlugin.Set(&Customer1.Orders[0].Key, Customer1.Orders[0].Content)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(&Customer1.Orders[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Orders[0].Content))
-			})
-		})
-		When("Valid Mutliple Sub Collection Set", func() {
-			It("Should store item successfully", func() {
-				err := docPlugin.Set(&Customer1.Reviews[0].Key, Customer1.Reviews[0].Content)
-				Expect(err).ShouldNot(HaveOccurred())
-				
-				doc, err := docPlugin.Get(&Customer1.Reviews[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Reviews[0].Content))
-			})
-		})
-	})
-}
-
-func DeleteTests(docPlugin document.DocumentService) {
-	Context("Delete", func() {
-		When("Blank key.Collection.Name", func() {
-			It("Should return error", func() {
-				key := document.Key{Id: "1"}
-				err := docPlugin.Delete(&key)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Blank key.Id", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}}
-				err := docPlugin.Delete(&key)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid Delete", func() {
-			It("Should delete item successfully", func() {
-				docPlugin.Set(&UserKey1, UserItem1)
-
-				err := docPlugin.Delete(&UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(&UserKey1)
-				Expect(doc).To(BeNil())
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid Sub Collection Delete", func() {
-			It("Should delete item successfully", func() {
-				docPlugin.Set(&Customer1.Orders[0].Key, Customer1.Orders[0].Content)
-
-				err := docPlugin.Delete(&Customer1.Orders[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(&Customer1.Orders[0].Key)
-				Expect(doc).To(BeNil())
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid Parent and Sub Collection Delete", func() {
-			It("Should delete all children", func() {
-				LoadCustomersData(docPlugin)
-
-				col := document.Collection{
-					Name: "orders",
-					Parent: &document.Key{
-						Collection: &document.Collection{
-							Name: "customers",
-						},
-					},
-				}
-
-				result, err := docPlugin.Query(&col, []document.QueryExpression{}, 0, nil)
-				Expect(err).To(BeNil())
-				Expect(result.Documents).To(HaveLen(5))
-
-				err = docPlugin.Delete(&Customer1.Key)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				err = docPlugin.Delete(&Customer2.Key)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				result, err = docPlugin.Query(&col, []document.QueryExpression{}, 0, nil)
-				Expect(err).To(BeNil())
-				Expect(result.Documents).To(HaveLen(0))
-			})
-		})
-	})
-}
-
-func QueryTests(docPlugin document.DocumentService) {
-	Context("Query", func() {
-		When("Invalid - blank key.Collection.Name", func() {
-			It("Should return an error", func() {
-				result, err := docPlugin.Query(&document.Collection{}, []document.QueryExpression{}, 0, nil)
-				Expect(result).To(BeNil())
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Invalid - nil expressions argument", func() {
-			It("Should return an error", func() {
-				result, err := docPlugin.Query(&document.Collection{Name: "users"}, nil, 0, nil)
-				Expect(result).To(BeNil())
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Empty database", func() {
-			It("Should return empty list", func() {
-				result, err := docPlugin.Query(&document.Collection{Name: "users"}, []document.QueryExpression{}, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-				Expect(result.PagingToken).To(BeNil())
-			})
-		})
-		When("key: {users}, subcol: '', exp: []", func() {
-			It("Should return all users", func() {
-				LoadUsersData(docPlugin)
-				LoadCustomersData(docPlugin)
-
-				result, err := docPlugin.Query(&document.Collection{Name: "users"}, []document.QueryExpression{}, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(3))
-
-				for _, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					Expect(d.Key.Collection.Name).To(Equal("users"))
-					Expect(d.Key.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Parent).To(BeNil())
-				}
-			})
-		})
-		When("key: {customers, nil}, subcol: '', exp: []", func() {
-			It("Should return 2 items", func() {
-				LoadCustomersData(docPlugin)
-
-				result, err := docPlugin.Query(&CustomersColl, []document.QueryExpression{}, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.Documents[0].Content["email"]).To(BeEquivalentTo(Customer1.Content["email"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[1].Content["email"]).To(BeEquivalentTo(Customer2.Content["email"]))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer2.Key))
-				Expect(result.PagingToken).To(BeNil())
-			})
-		})
-		When("key: {customers, nil}, subcol: '', exp: [country == US]", func() {
-			It("Should return 1 item", func() {
-				LoadCustomersData(docPlugin)
-
-				exps := []document.QueryExpression{
-					{Operand: "country", Operator: "==", Value: "US"},
-				}
-				result, err := docPlugin.Query(&CustomersColl, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(1))
-				Expect(result.Documents[0].Content["email"]).To(BeEquivalentTo(Customer2.Content["email"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer2.Key))
-				Expect(result.PagingToken).To(BeNil())
-			})
-		})
-		When("key: {customers, nil}, subcol: '', exp: [country == US, age > 40]", func() {
-			It("Should return 0 item", func() {
-				LoadCustomersData(docPlugin)
-
-				exps := []document.QueryExpression{
-					{Operand: "country", Operator: "==", Value: "US"},
-					{Operand: "age", Operator: ">", Value: "40"},
-				}
-				result, err := docPlugin.Query(&CustomersColl, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-			})
-		})
-		When("key: {customers, key1}, subcol: orders", func() {
-			It("Should return 3 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				result, err := docPlugin.Query(&coll, []document.QueryExpression{}, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(3))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[0].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[0].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[1].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[1].Content["testName"]))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer1.Orders[1].Key))
-				Expect(*result.Documents[1].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[2].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[2].Content["testName"]))
-				Expect(*result.Documents[2].Key).To(BeEquivalentTo(Customer1.Orders[2].Key))
-				Expect(*result.Documents[2].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-			})
-		})
-		When("key: {customers, nil}, subcol: orders, exps: [number == 1]", func() {
-			It("Should return 2 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "==", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[0].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[0].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[1].Content["testName"]).To(BeEquivalentTo(Customer2.Orders[0].Content["testName"]))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer2.Orders[0].Key))
-				Expect(*result.Documents[1].Key.Collection.Parent).To(BeEquivalentTo(Customer2.Key))
-			})
-		})
-		When("key: {customers, key1}, subcol: orders, exps: [number == 1]", func() {
-			It("Should return 1 order", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "==", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(1))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[0].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[0].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-			})
-		})
-		When("key: {customers, nil}, subcol: orders, exps: [number > 1]", func() {
-			It("Should return 3 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: ">", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(3))
-
-				for _, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					Expect(d.Key.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Name).To(Equal("orders"))
-					Expect(d.Key.Collection.Parent).ToNot(BeNil())
-					Expect(d.Key.Collection.Parent.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Parent.Collection.Name).To(Equal("customers"))
-				}
-			})
-		})
-		When("key: {customers, key1}, subcol: orders, exps: [number > 1]", func() {
-			It("Should return 2 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: ">", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[1].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[1].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[1].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[2].Content["testName"]))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer1.Orders[2].Key))
-				Expect(*result.Documents[1].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-			})
-		})
-		When("key: {customers, nil}, subcol: orders, exps: [number < 1]", func() {
-			It("Should return 2 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "<", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-
-				for _, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					Expect(d.Key.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Name).To(Equal("orders"))
-					Expect(d.Key.Collection.Parent).ToNot(BeNil())
-					Expect(d.Key.Collection.Parent.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Parent.Collection.Name).To(Equal("customers"))
-				}
-			})
-		})
-		When("key: {customers, key1}, subcol: orders, exps: [number < 1]", func() {
-			It("Should return 0 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "<", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-			})
-		})
-		When("key: {customers, nil}, subcol: orders, exps: [number >= 1]", func() {
-			It("Should return 5 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: ">=", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(5))
-
-				for _, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					Expect(d.Key.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Name).To(Equal("orders"))
-					Expect(d.Key.Collection.Parent).ToNot(BeNil())
-					Expect(d.Key.Collection.Parent.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Parent.Collection.Name).To(Equal("customers"))
-				}
-			})
-		})
-		When("key: {customers, key1}, subcol: orders, exps: [number >= 1]", func() {
-			It("Should return 3 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: ">=", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(3))
-
-				for _, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					Expect(d.Key.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Name).To(Equal("orders"))
-					Expect(d.Key.Collection.Parent).ToNot(BeNil())
-					Expect(d.Key.Collection.Parent.Id).ToNot(Equal(""))
-					Expect(d.Key.Collection.Parent.Collection.Name).To(Equal("customers"))
-				}
-			})
-		})
-		When("key: {customers, nil}, subcol: orders, exps: [number <= 1]", func() {
-			It("Should return 2 orders", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "<=", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[0].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[0].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(result.Documents[1].Content["testName"]).To(BeEquivalentTo(Customer2.Orders[0].Content["testName"]))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer2.Orders[0].Key))
-				Expect(*result.Documents[1].Key.Collection.Parent).To(BeEquivalentTo(Customer2.Key))
-			})
-		})
-		When("key: {customers, key1}, subcol: orders, exps: [number <= 1]", func() {
-			It("Should return 1 order", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "number", Operator: "<=", Value: "1"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(1))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[0].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[0].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-			})
-		})
-		When("key {customers, nil}, subcol: orders, exps: [type startsWith scooter]", func() {
-			It("Should return 2 order", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &CustomersKey,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "type", Operator: "startsWith", Value: "scooter"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[2].Content["testName"]))
-				Expect(result.Documents[1].Content["testName"]).To(BeEquivalentTo(Customer2.Orders[1].Content["testName"]))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[2].Key))
-				Expect(*result.Documents[1].Key.Collection.Parent).To(BeEquivalentTo(Customer2.Key))
-				Expect(*result.Documents[1].Key).To(BeEquivalentTo(Customer2.Orders[1].Key))
-			})
-		})
-		When("key {customers, key1}, subcol: orders, exps: [type startsWith bike/road]", func() {
-			It("Should return 1 order", func() {
-				LoadCustomersData(docPlugin)
-
-				coll := document.Collection{
-					Name:   "orders",
-					Parent: &Customer1.Key,
-				}
-				exps := []document.QueryExpression{
-					{Operand: "type", Operator: "startsWith", Value: "scooter"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 0, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(1))
-				Expect(result.Documents[0].Content["testName"]).To(BeEquivalentTo(Customer1.Orders[2].Content["testName"]))
-				Expect(*result.Documents[0].Key).To(BeEquivalentTo(Customer1.Orders[2].Key))
-				Expect(*result.Documents[0].Key.Collection.Parent).To(BeEquivalentTo(Customer1.Key))
-			})
-		})
-		When("key: {items, nil}, subcol: '', exp: [], limit: 10", func() {
-			It("Should return have multiple pages", func() {
-				LoadItemsData(docPlugin)
-
-				coll := document.Collection{
-					Name: "items",
-				}
-				result, err := docPlugin.Query(&coll, []document.QueryExpression{}, 10, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(10))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				dataMap := make(map[string]string)
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					dataMap[val] = val
-				}
-
-				result, err = docPlugin.Query(&coll, []document.QueryExpression{}, 10, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.PagingToken).To(BeNil())
-
-				// Ensure values are unique
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					if _, found := dataMap[val]; found {
-						Expect("matching value").ShouldNot(HaveOccurred())
-					}
-				}
-			})
-		})
-		When("key: {items, nil}, subcol: '', exps: [letter > D], limit: 4", func() {
-			It("Should return have multiple pages", func() {
-				LoadItemsData(docPlugin)
-
-				coll := document.Collection{
-					Name: "items",
-				}
-				exps := []document.QueryExpression{
-					{Operand: "letter", Operator: ">", Value: "D"},
-				}
-				result, err := docPlugin.Query(&coll, exps, 4, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(4))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				dataMap := make(map[string]string)
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					dataMap[val] = val
-				}
-
-				result, err = docPlugin.Query(&coll, exps, 4, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(4))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					if _, found := dataMap[val]; found {
-						Expect("matching value").ShouldNot(HaveOccurred())
-					}
-				}
-
-				result, err = docPlugin.Query(&coll, exps, 4, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-				Expect(result.PagingToken).To(BeEmpty())
-			})
-		})
-		When("key: {parentItems, 1}, subcol: items, exp: [], limit: 10", func() {
-			It("Should return have multiple pages", func() {
-				LoadItemsData(docPlugin)
-
-				result, err := docPlugin.Query(&ChildItemsCollection, []document.QueryExpression{}, 10, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(10))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				dataMap := make(map[string]string)
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					dataMap[val] = val
-				}
-
-				result, err = docPlugin.Query(&ChildItemsCollection, []document.QueryExpression{}, 10, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(2))
-				Expect(result.PagingToken).To(BeNil())
-
-				// Ensure values are unique
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					if _, found := dataMap[val]; found {
-						Expect("matching value").ShouldNot(HaveOccurred())
-					}
-				}
-			})
-		})
-		When("key: {parentItems, 1}, subcol: items, exps: [letter > D], limit: 4", func() {
-			It("Should return have multiple pages", func() {
-				LoadItemsData(docPlugin)
-
-				exps := []document.QueryExpression{
-					{Operand: "letter", Operator: ">", Value: "D"},
-				}
-				result, err := docPlugin.Query(&ChildItemsCollection, exps, 4, nil)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(4))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				dataMap := make(map[string]string)
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					dataMap[val] = val
-				}
-
-				result, err = docPlugin.Query(&ChildItemsCollection, exps, 4, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(4))
-				Expect(result.PagingToken).ToNot(BeEmpty())
-
-				// Ensure values are unique
-				for i, d := range result.Documents {
-					Expect(d.Key).ToNot(BeNil())
-					val := fmt.Sprintf("%v", result.Documents[i].Content["letter"])
-					if _, found := dataMap[val]; found {
-						Expect("matching value").ShouldNot(HaveOccurred())
-					}
-				}
-
-				result, err = docPlugin.Query(&ChildItemsCollection, exps, 4, result.PagingToken)
-				Expect(result).ToNot(BeNil())
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(result.Documents).To(HaveLen(0))
-				Expect(result.PagingToken).To(BeEmpty())
-			})
-		})
-	})
-}
